Adding sql file
