+++ server/db.sql
@@ -1,0 +1,4 @@
+-- SQLite
+
+SELECT * from users;
+SELECT * from universities;
