+++ server/main.go
@@ -16,6 +16,15 @@ type University struct {
 	Country string `json:"country"`
 }
 
+type User struct {
+	gorm.Model
+	ID        uint   `gorm:"primaryKey;autoIncrement" json:"User ID"`
+	FirstName string `json:"First Name"`
+	LastName  string `json:"Last Name"`
+	Email     string `json:"username"`
+	Password  string `json:"password"`
+}
+
 type OfferedCourse struct {
 	gorm.Model
 	ID uint `gorm:"primaryKey;autoIncrement" json:"Course ID"`
@@ -34,6 +43,19 @@ func getUniversities(c *gin.Context) {
 
 }
 
+func getUsers(c *gin.Context) {
+	db, err := gorm.Open(sqlite.Open("unigenie.db"), &gorm.Config{})
+	if err != nil {
+		panic("failed to connect database")
+	}
+
+	users := []User{}
+
+	db.Find(&users)
+	c.JSON(200, &users)
+
+}
+
 func setupRouter() *gin.Engine {
 	// Disable Console Color
 	// gin.DisableConsoleColor()
@@ -49,8 +71,8 @@ func setupRouter() *gin.Engine {
 func main() {
 	r := setupRouter()
 	setDatabase()
-	r.GET("/universities", getUniversities)
-
+	r.GET("/getUniversities", getUniversities)
+	r.GET("/getUsers", getUsers)
 	// Listen and Server in 0.0.0.0:8080
 	r.Run(":8080")
 }
